+++ packages/platforms/react-native/__mocks__/react-native.ts
@@ -4,6 +4,8 @@ type PlatformConstants
   = { Manufacturer: string, Model: string, Release: string }
   | Record<string, never>
 
+const VALID_API_KEY = 'a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6'
+
 export const Platform = new class {
   private os: Os = 'ios'
 
@@ -105,35 +107,82 @@ const createPool = () => {
   return '0123456789abcdef'.repeat(128)
 }
 
+const getDevice = (): DeviceInfo => {
+  switch (Platform.OS) {
+    case 'ios':
+      return {
+        arch: 'arm64',
+        model: 'iPhone14,1',
+        bundleVersion: '12345',
+        bundleIdentifier: 'my.cool.app'
+      }
+
+    case 'android':
+      return {
+        arch: 'x86',
+        model: 'TheGoodPhone1',
+        versionCode: '6789',
+        bundleIdentifier: 'my.cool.app'
+      }
+
+    case 'other':
+      return {}
+  }
+}
+
 const BugsnagReactNativePerformance = {
-  getDeviceInfo (): DeviceInfo {
-    switch (Platform.OS) {
-      case 'ios':
-        return {
-          arch: 'arm64',
-          model: 'iPhone14,1',
-          bundleVersion: '12345',
-          bundleIdentifier: 'my.cool.app'
-        }
-
-      case 'android':
-        return {
-          arch: 'x86',
-          model: 'TheGoodPhone1',
-          versionCode: '6789',
-          bundleIdentifier: 'my.cool.app'
-        }
-
-      case 'other':
-        return {}
-    }
-  },
+  getDeviceInfo: jest.fn(getDevice),
   requestEntropy: jest.fn(() => {
     return createPool()
   }),
   requestEntropyAsync: jest.fn(() => {
     return Promise.resolve(createPool())
-  })
+  }),
+  isNativePerformanceAvailable: jest.fn(() => {
+    return true
+  }),
+  attachToNativeSDK: jest.fn(() => {
+    return {
+      apiKey: VALID_API_KEY,
+      endpoint: '/traces',
+      appVersion: '1.2.3',
+      releaseStage: 'test',
+      enabledReleaseStages: ['test', 'production'],
+      serviceName: 'com.native.performance.test',
+      attributeCountLimit: 128,
+      attributeStringValueLimit: 1024,
+      attributeArrayLengthLimit: 1000
+    }
+  }),
+  startNativeSpan: jest.fn((name, options) => {
+    return {
+      name,
+      id: 'native-span-id',
+      traceId: options.parentContext?.traceId || 'native-trace-id',
+      startTime: options.startTime,
+      parentSpanId: options.parentContext?.id || undefined
+    }
+  }),
+  endNativeSpan: jest.fn(() => {
+    return Promise.resolve()
+  }),
+  markNativeSpanEndTime: jest.fn(),
+  discardNativeSpan: jest.fn(() => {
+    return Promise.resolve()
+  }),
+  getNativeConstants () {
+    return {
+      CacheDir: '/mock/CacheDir',
+      DocumentDir: '/mock/DocumentDir'
+    }
+  },
+  exists: jest.fn(),
+  isDir: jest.fn(),
+  ls: jest.fn(),
+  mkdir: jest.fn(),
+  readFile: jest.fn(),
+  unlink: jest.fn(),
+  writeFile: jest.fn()
 }
 
 export const TurboModuleRegistry = {
@@ -146,6 +195,10 @@ export const TurboModuleRegistry = {
         return null
     }
   }
+}
+
+export const NativeModules = {
+  BugsnagReactNativePerformance
 }
 
 export type AppStateStatus = 'active' | 'inactive' | 'background'
@@ -171,3 +224,7 @@ export const AppState = new class {
     }
   }
 }()
+
+export const AppRegistry = {
+  setWrapperComponentProvider: jest.fn()
+}