+++ packages/plugin-react-native-navigation/lib/react-native-navigation-plugin.ts
@@ -1,8 +1,6 @@
 import type { Plugin, SpanFactory, SpanInternal } from '@bugsnag/core-performance'
-import type { ReactNativeConfiguration } from '@bugsnag/react-native-performance'
+import type { ReactNativeConfiguration, ReactNativeSpanFactory } from '@bugsnag/react-native-performance'
 import type { NavigationDelegate } from 'react-native-navigation/lib/dist/src/NavigationDelegate'
-
-import { createNavigationSpan } from '@bugsnag/react-native-performance'
 
 type Reason = 'immediate' | 'mount' | 'unmount' | 'condition'
 
@@ -16,7 +14,7 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
   private startTimeout?: NodeJS.Timeout
   private endTimeout?: NodeJS.Timeout
   private componentsWaiting = 0
-  private spanFactory?: SpanFactory<ReactNativeConfiguration>
+  private spanFactory?: ReactNativeSpanFactory
   private previousRoute?: string
 
   constructor (Navigation: NavigationDelegate) {
@@ -67,7 +65,7 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
   }
 
   configure (configuration: ReactNativeConfiguration, spanFactory: SpanFactory<ReactNativeConfiguration>) {
-    this.spanFactory = spanFactory
+    this.spanFactory = spanFactory as ReactNativeSpanFactory
 
     // Potential for a navigation to occur
     this.Navigation.events().registerCommandListener((name, params) => {
@@ -80,11 +78,11 @@ class BugsnagPluginReactNativeNavigationPerformance implements Plugin<ReactNativ
 
     // Navigation has occurred
     this.Navigation.events().registerComponentWillAppearListener(event => {
-      if (typeof this.startTime === 'number') {
+      if (this.spanFactory && typeof this.startTime === 'number') {
         clearTimeout(this.startTimeout)
 
         const routeName = event.componentName
-        this.currentNavigationSpan = createNavigationSpan(spanFactory, routeName, { startTime: this.startTime })
+        this.currentNavigationSpan = this.spanFactory.startNavigationSpan(routeName, { startTime: this.startTime, doNotDelegateToNativeSDK: true })
         this.currentNavigationSpan.setAttribute('bugsnag.navigation.triggered_by', '@bugsnag/plugin-react-native-navigation-performance')
 
         if (this.previousRoute) {