+++ test/browser/features/fixtures/packages/resource-load-spans/src/index.js
@@ -12,7 +12,7 @@ document.getElementById("end-span").addEventListener("click", () => {
     const node = document.getElementById("image-container")
     const img = new Image()
     node.appendChild(img)
-    img.src = "/favicon.png?height=100&width=100"
+    img.src = "/docs/favicon.png?height=100&width=100"
     img.onload = () => {
         span.end()
     }